Rename config.py to qa_config.py
